Remove volatility from schema manifest file (#160)
* Remove volatility from schema manifest file

The schema paths written to the schema manifest are full absolute paths.
This introduces volatility in the build, makes it non-reproducible and
non-relocatable, since the full absolute path to a schema will vary
between machines, e.g., between a local user and a CI machine. An
additional source of volatility is the timestamp written to the file
which will vary between every invocation even on the same machine.

With this change, the schema paths written to the schema manifest are
relativized to the base project directory. The relative path to a schema
is all that's required locate it and will always be consistent between
machines. Additionally, the java-ordered-properties library is used to
ensure the properties file is always written in a deterministic order
and removes the unnecessary timestamp.

* Do not write schema-manifest.props file to sources directory